move util function tests to test_util
